Rearrange concentration and fuzzy
All files in the old lib/ folder are moved out.
