+++ chart_of_intervals.py
@@ -1,10 +1,9 @@
 from typing import List
 
-import fuzzy.parse as parse
-from fuzzy.classes import Interval
-from lib.concentration_grader import save_chart_of_intervals
+import concentration.fuzzy.parse as parse
+from concentration.fuzzy.classes import Interval
 
 
 if __name__ == "__main__":
     intervals: List[Interval] = parse.read_intervals_from_json("intervals.json")
-    save_chart_of_intervals("intervals.jpg", intervals)
+    parse.save_chart_of_intervals("intervals.jpg", intervals)
