+++ intergrated_gui/gui_state.ini
@@ -2,19 +2,19 @@
 checked = True
 camera_dist = 45
 warn_dist = 40
-warning_enabled = True
+warning_enabled = False
 
 [time]
 checked = True
-time_limit = 10
-break_time = 3
-warning_enabled = True
+time_limit = 30
+break_time = 10
+warning_enabled = False
 
 [posture]
 checked = True
 warn_angle = strict
 model_path = default
-warning_enabled = True
+warning_enabled = False
 
 [brightness]
 checked = False
